Check if any log exists

+++ catLogs.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-for l in $(ls -rt "$1"/api.*.log); do
-	cat "$l" >> "$1/api.log"
-	rm -f "$l"
-done
+if [ -e "$1"/api.*.log ]; then
+	for l in $(ls -rt "$1"/api.*.log); do
+		echo Archiving "$l"...
+		cat "$l" >> "$1/api.log"
+		rm -f "$l"
+	done
+fi
